refactor(units): create units folder (#4)
