+++ CMakeLists.txt
@@ -84,16 +84,24 @@ find_program(POSU_CLANG_TIDY   clang-tidy)
 ##################
 
 set(
+    POSU_UNITS_HPP_FILES
+        ${CMAKE_CURRENT_LIST_DIR}/posu/units/base_unit.hpp
+        ${CMAKE_CURRENT_LIST_DIR}/posu/units/units.hpp
+)
+set(
+    POSU_UNITS_IPP_FILES
+        ${CMAKE_CURRENT_LIST_DIR}/posu/units/ipp/base_unit.ipp
+        ${CMAKE_CURRENT_LIST_DIR}/posu/units/ipp/units.ipp
+)
+
+set(
     POSU_HPP_FILES
-        ${CMAKE_CURRENT_LIST_DIR}/posu/units.hpp
         ${CMAKE_CURRENT_LIST_DIR}/posu/utility.hpp
         ${CMAKE_CURRENT_LIST_DIR}/posu/vmath.hpp
         ${CMAKE_CURRENT_LIST_DIR}/posu/zip_range.hpp
 )
-
 set(
     POSU_IPP_FILES
-        ${CMAKE_CURRENT_LIST_DIR}/posu/ipp/units.ipp
         ${CMAKE_CURRENT_LIST_DIR}/posu/ipp/utility.ipp
         ${CMAKE_CURRENT_LIST_DIR}/posu/ipp/vmath.ipp
         ${CMAKE_CURRENT_LIST_DIR}/posu/ipp/zip_range.ipp
@@ -121,7 +129,7 @@ if(BUILD_TESTING AND POSU_ENABLE_TESTING)
     add_test(NAME test_zip_range COMMAND test_zip_range)
     set_tests_properties(test_zip_range PROPERTIES DEPENDS test_arith_tuple)
 
-    add_executable(test_units ${CMAKE_CURRENT_LIST_DIR}/test/test_units.cpp)
+    add_executable(test_units ${CMAKE_CURRENT_LIST_DIR}/test/units/test_units.cpp)
     target_link_libraries(test_units PRIVATE Catch2::Catch2)
     target_link_libraries(test_units PRIVATE posu)
     add_test(NAME test_units COMMAND test_units)
@@ -141,8 +149,8 @@ if(POSU_CLANG_FORMAT)
                     ${POSU_HPP_FILES}
                     ${POSU_IPP_FILES}
                     ${CMAKE_CURRENT_LIST_DIR}/test/test_arith_tuple.cpp
-                    ${CMAKE_CURRENT_LIST_DIR}/test/test_units.cpp
                     ${CMAKE_CURRENT_LIST_DIR}/test/test_zip_range.cpp
+                    ${CMAKE_CURRENT_LIST_DIR}/test/units/test_units.cpp
             DEPENDS
                 ${POSU_HPP_FILES}
     )
@@ -164,13 +172,13 @@ if(POSU_CLANG_TIDY)
                     -p ${CMAKE_CURRENT_BINARY_DIR}
                     ${POSU_HPP_FILES}
                     ${CMAKE_CURRENT_LIST_DIR}/test/test_arith_tuple.cpp
-                    ${CMAKE_CURRENT_LIST_DIR}/test/test_units.cpp
                     ${CMAKE_CURRENT_LIST_DIR}/test/test_zip_range.cpp
+                    ${CMAKE_CURRENT_LIST_DIR}/test/units/test_units.cpp
                 DEPENDS
                     ${POSU_HPP_FILES}
                     ${CMAKE_CURRENT_LIST_DIR}/test/test_arith_tuple.cpp
-                    ${CMAKE_CURRENT_LIST_DIR}/test/test_units.cpp
                     ${CMAKE_CURRENT_LIST_DIR}/test/test_zip_range.cpp
+                    ${CMAKE_CURRENT_LIST_DIR}/test/units/test_units.cpp
     )
 else()
     message(STATUS "clang-tidy not found")
